+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsRoleMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsRoleMapper {
-    int countByExample(UmsRoleExample example);
+    long countByExample(UmsRoleExample example);
 
     int deleteByExample(UmsRoleExample example);
